+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [binjr-adapter-demo 3.0.0](https://github.com/binjr/binjr/releases/tag/3.0.0)
+Released on Fri, 16 Apr 2021
+
+* _[New]_ Plugin compatible with binjr v3.0.0 and higher
+* _[New]_ Added a demo log file source.
+
 ## [binjr-adapter-demo 1.1.0](https://github.com/binjr/binjr/releases/tag/1.1.0)
 Released on Thu, 24 Oct 2019